requirements.txt
django==5.1.5
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.4.0
pillow==10.3.0
channels==4.1.0
daphne==4.1.0
python-dotenv==1.0.1